refactor: repository interface를 application layer로 이관
