+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/feed/query/FeedQueryRepositoryImpl.java
@@ -1,5 +1,6 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.feed.query;
 
+import com.LetMeDoWith.LetMeDoWith.application.feed.repository.FeedQueryRepository;
 import com.LetMeDoWith.LetMeDoWith.common.enums.common.Yn;
 import com.LetMeDoWith.LetMeDoWith.common.util.SystemTimeUtil;
 import com.LetMeDoWith.LetMeDoWith.domain.feedback.model.QDowithTaskFeedback;
@@ -33,7 +34,8 @@ public class FeedQueryRepositoryImpl implements FeedQueryRepository {
     private final QBadge badge = QBadge.badge;
 
     @Override
-    public List<FeedDowithTaskQueryDto> getFeedbackAvailableDowithTasks(LocalDateTime referenceDateTime) {
+    public List<FeedDowithTaskQueryDto> getFeedbackAvailableDowithTasks(
+        LocalDateTime referenceDateTime) {
 
         LocalDateTime now = referenceDateTime != null ? referenceDateTime : SystemTimeUtil.now();
 
@@ -50,38 +52,39 @@ public class FeedQueryRepositoryImpl implements FeedQueryRepository {
         if (today.equals(startRangeDate)) {
             // Case 1: 범위 시작과 끝이 같은 날짜인 경우
             condition.and(dowithTask
-                    .date
-                    .eq(today)
-                    .and(dowithTask.startTime.gt(startRangeTime))
-                    .and(dowithTask.startTime.loe(nowTime)));
+                .date
+                .eq(today)
+                .and(dowithTask.startTime.gt(startRangeTime))
+                .and(dowithTask.startTime.loe(nowTime)));
         } else {
             // Case 2: 날짜가 걸쳐있는 경우 (자정 직후)
-            condition.and((dowithTask.date.eq(startRangeDate).and(dowithTask.startTime.gt(startRangeTime)))
+            condition.and(
+                (dowithTask.date.eq(startRangeDate).and(dowithTask.startTime.gt(startRangeTime)))
                     .or(dowithTask.date.eq(today).and(dowithTask.startTime.loe(nowTime))));
         }
 
         return queryFactory
-                .select(Projections.constructor(
-                        FeedDowithTaskQueryDto.class,
-                        dowithTask.id,
-                        dowithTask.memberId,
-                        member.nickname,
-                        badge.imageUrl,
-                        dowithTask.title,
-                        dowithTask.status.stringValue(),
-                        dowithTask.date,
-                        dowithTask.startTime,
-                        JPAExpressions.select(dowithTaskFeedback.count())
-                                .from(dowithTaskFeedback)
-                                .where(dowithTaskFeedback.dowithTaskId.eq(dowithTask.id))))
-                .from(dowithTask)
-                .leftJoin(member)
-                .on(member.id.eq(dowithTask.memberId))
-                .leftJoin(memberBadge)
-                .on(memberBadge.memberId.eq(member.id).and(memberBadge.isMain.eq(Yn.TRUE)))
-                .leftJoin(badge)
-                .on(badge.id.eq(memberBadge.badge.id))
-                .where(condition)
-                .fetch();
+            .select(Projections.constructor(
+                FeedDowithTaskQueryDto.class,
+                dowithTask.id,
+                dowithTask.memberId,
+                member.nickname,
+                badge.imageUrl,
+                dowithTask.title,
+                dowithTask.status.stringValue(),
+                dowithTask.date,
+                dowithTask.startTime,
+                JPAExpressions.select(dowithTaskFeedback.count())
+                    .from(dowithTaskFeedback)
+                    .where(dowithTaskFeedback.dowithTaskId.eq(dowithTask.id))))
+            .from(dowithTask)
+            .leftJoin(member)
+            .on(member.id.eq(dowithTask.memberId))
+            .leftJoin(memberBadge)
+            .on(memberBadge.memberId.eq(member.id).and(memberBadge.isMain.eq(Yn.TRUE)))
+            .leftJoin(badge)
+            .on(badge.id.eq(memberBadge.badge.id))
+            .where(condition)
+            .fetch();
     }
 }
